Improve memory usage
git-svn-id: file:///home/svn/DVA2006@11323 1e6d2d28-e1fb-0310-b8ab-81908d8c6b6a

+++ src/test/org/deltava/acars/PlotLocalMap.java
@@ -3,6 +3,7 @@ package org.deltava.acars;
 import java.sql.*;
 import java.util.*;
 import java.util.concurrent.*;
+import java.util.stream.Collectors;
 
 import org.deltava.beans.schedule.Airport;
 import org.deltava.util.system.SystemData;
@@ -11,20 +12,16 @@ public class PlotLocalMap extends PlotMap {
 	
 	private static final int MIN_ZOOM = 10;
 	private static final int MAX_ZOOM = 13;
-	private static final int MAX_TILES = 325 * 1024;
+	private static final int MAX_TILES = 320 * 1024;
 	
-	private static final List<String> _airports = Arrays.asList("ATL", "YYZ", "MIA", "NRT",  "DFW", "IAD", "HKX", "HKG", "HNL", "OGG", "YUL", "LHR", "ICN", "BDA", "YVR", "SYD", "ZRH", "PPT", "ANC", "GVA"); 
-	
-	@Override
-	protected void setUp() throws Exception {
-		super.setUp();
+	PlotLocalMap() {
 		_zooms.put(Integer.valueOf(10), new ProjectInfo(10, 64, 10));
-		_zooms.put(Integer.valueOf(11), new ProjectInfo(11, 40, 8));
-		_zooms.put(Integer.valueOf(12), new ProjectInfo(12, 28, 6));
-		_zooms.put(Integer.valueOf(13), new ProjectInfo(13, 20, 4));
+		_zooms.put(Integer.valueOf(11), new ProjectInfo(11, 32, 6));
+		_zooms.put(Integer.valueOf(12), new ProjectInfo(12, 24, 3));
+		_zooms.put(Integer.valueOf(13), new ProjectInfo(13, 16, 2));
 	}
 	
-	protected static Collection<String> getPopularAirports(int max) throws Exception {
+	protected static Collection<String> getPopularAirports(int cnt, int st) throws Exception {
 		try (Connection c = DriverManager.getConnection(URL)) {
 			c.setAutoCommit(false);
 			try (Statement s = c.createStatement()) {
@@ -37,8 +34,8 @@ public class PlotLocalMap extends PlotMap {
 			Collection<String> aps = new LinkedHashSet<String>();
 			try (Statement s = c.createStatement()) {
 				StringBuilder sqlBuf = new StringBuilder("SELECT IATA FROM AP_COUNT ORDER BY CNT DESC");
-				if (max > 0)
-					sqlBuf.append(" LIMIT ").append(max);
+				if (cnt > 0)
+					sqlBuf.append(" LIMIT ").append(st).append(',').append(cnt);
 				
 				try (ResultSet rs = s.executeQuery(sqlBuf.toString())) {
 					while (rs.next())
@@ -47,12 +44,14 @@ public class PlotLocalMap extends PlotMap {
 			}
 			
 			// Write the track airport count
-			try (Statement s = c.createStatement()) {
-				s.execute("TRUNCATE TRACK_AIRPORTS");
+			if (st == 0) {
+				try (Statement s = c.createStatement()) {
+					s.execute("TRUNCATE TRACK_AIRPORTS");
+				}
 			}
 			
-			Collection<String> allAP = new LinkedHashSet<String>(aps); allAP.addAll(_airports);
-			StringBuilder buf = new StringBuilder("INSERT INTO TRACK_AIRPORTS (SELECT IATA, CNT FROM AP_COUNT WHERE IATA IN (");
+			Collection<String> allAP = new LinkedHashSet<String>(aps);
+			StringBuilder buf = new StringBuilder("REPLACE INTO TRACK_AIRPORTS (SELECT IATA, CNT FROM AP_COUNT WHERE IATA IN (");
 			for (Iterator<?> i = allAP.iterator(); i.hasNext(); ) {
 				i.next(); buf.append('?');
 				if (i.hasNext())
@@ -73,18 +72,13 @@ public class PlotLocalMap extends PlotMap {
 		}
 	}
 	
-	public void testPlotMap() throws Exception {
+	public void plotMap(int start, int count) throws Exception {
 		
 		// Init the filter
-		Collection<String> apCodes = getPopularAirports(70);
-		apCodes.addAll(_airports);
-
-		Collection<Airport> airports = new ArrayList<Airport>();
-		for (String iata : apCodes)
-			airports.add(SystemData.getAirport(iata));
-		
-		AirportEntryFilter f = new AirportEntryFilter(MAX_ZOOM, 110, airports);
-		f.setMaxAltitude(18500);
+		Collection<String> apCodes = getPopularAirports(count, start);
+		Collection<Airport> airports = apCodes.stream().map(code -> SystemData.getAirport(code)).filter(Objects::nonNull).collect(Collectors.toList());
+		AirportEntryFilter f = new AirportEntryFilter(MAX_ZOOM, 95, airports);
+		f.setMaxAltitude(16900);
 		
 		// Load flight IDs
 		log.info("Loading Flight IDs");
@@ -95,7 +89,8 @@ public class PlotLocalMap extends PlotMap {
 		loadFlights(IDs, gt, f);
 		
 		// Truncate tracks
-		truncateTracks(MIN_ZOOM);
+		if (start == 0)
+			truncateTracks(MIN_ZOOM);
 		
 		while (gt.getZoom() >= MIN_ZOOM) {
 			SparseGlobalTile pgt = scale(gt);
